implemented average damage calculation for multi-hit-moves, added helper fucntions for setting up party programmatically, updated readme, added more clarity for screenshot pokemon determination

+++ src/pokemon.go
@@ -46,7 +46,7 @@ type Move struct {
 	Rank     int     `json:"rank"`
 	Name     string  `json:"name"`
 	Type     string  `json:"type"`
-	Damage   int     `json:"damage"`
+	Damage   float64 `json:"damage"`
 	Accuracy float64 `json:"accuracy"`
 }
